+++ CHANGELOG.md
@@ -8,6 +8,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
+
+### Changed
+
+### Fixed
+
+## [0.18.0] - 2026-01-31
+
+### Added
 - **Dialog Overlay System** - New reusable dialog component for displaying overlay content on top of views(when right panel focused) 
 - **League Standings Dialog** - Press `s` to view current league standings with highlighted match teams
 - **Match Formations Dialog** - Press `f` to view both teams' formations and starting lineups with player ratings